requirements.txt
langchain
langchain-ollama
langchain-chroma
pandas
streamlit
Flask
Pillow
